Add processor for ayumi
