--- libs/ayumi-js/README.md
@@ -1,8 +1,0 @@
-ayumi.js
-========
-
-Highly precise emulation of AY-3-8910 and YM2149 sound chips.
-
-Ported to JavaScript from [original Ayumi C code by Peter Sovietov](https://github.com/true-grue/ayumi).
-
-[Listen to it in action](http://pure-garden-1548.herokuapp.com/)
