+++ src/lib/stores/audio.ts
@@ -1,0 +1,3 @@
+import { writable } from 'svelte/store';
+
+export const audioContext = writable<AudioContext>();
