+++ src/routes/+layout.svelte
@@ -1,6 +1,22 @@
 <script lang="ts">
 	import DebugBar from '$lib/components/Debug/DebugBar.svelte';
+	import { onMount } from 'svelte';
 	import '../app.css';
+	import { audioContext } from '$lib/stores/audio';
+
+	let audioNode: AudioNode;
+
+	onMount(async () => {
+		if (!$audioContext) {
+			$audioContext = new AudioContext();
+			try {
+				await $audioContext.audioWorklet.addModule('/processor.js');
+			} catch (error) {}
+
+			audioNode = new AudioWorkletNode($audioContext, 'processor');
+			audioNode.connect($audioContext.destination);
+		}
+	});
 </script>
 
 <main class="h-full flex flex-col">
