+++ src/routes/+page.svelte
@@ -15,8 +15,9 @@
 	} from '$lib/stores/stores.js';
 	import ModuleEditor from '$lib/components/ModuleEditor/ModuleEditor.svelte';
 	import PlayerPlayFilled from '@tabler/icons-svelte/icons/player-play-filled';
-	import { playCurrentPattern } from '$lib/services/module-player';
 	import { isTrackPlaying } from '$lib/stores/debug';
+	import NoteData, { Note } from '$lib/models/note-data';
+	import { audioContext } from '$lib/stores/audio';
 
 	let fileLoaderInput: HTMLInputElement;
 	let unlisten: UnlistenFn;
@@ -67,17 +68,34 @@
 	}
 
 	async function playOrPauseCurrentPattern() {
+		$audioContext.resume();
+
+		const PT3ToneTable: number[] = [
+			0x0d10, 0x0c55, 0x0ba4, 0x0afc, 0x0a5f, 0x09ca, 0x093d, 0x08b8, 0x083b, 0x07c5, 0x0755,
+			0x06ec, 0x0688, 0x062a, 0x05d2, 0x057e, 0x052f, 0x04e5, 0x049e, 0x045c, 0x041d, 0x03e2,
+			0x03ab, 0x0376, 0x0344, 0x0315, 0x02e9, 0x02bf, 0x0298, 0x0272, 0x024f, 0x022e, 0x020f,
+			0x01f1, 0x01d5, 0x01bb, 0x01a2, 0x018b, 0x0174, 0x0160, 0x014c, 0x0139, 0x0128, 0x0117,
+			0x0107, 0x00f9, 0x00eb, 0x00dd, 0x00d1, 0x00c5, 0x00ba, 0x00b0, 0x00a6, 0x009d, 0x0094,
+			0x008c, 0x0084, 0x007c, 0x0075, 0x006f, 0x0069, 0x0063, 0x005d, 0x0058, 0x0053, 0x004e,
+			0x004a, 0x0046, 0x0042, 0x003e, 0x003b, 0x0037, 0x0034, 0x0031, 0x002f, 0x002c, 0x0029,
+			0x0027, 0x0025, 0x0023, 0x0021, 0x001f, 0x001d, 0x001c, 0x001a, 0x0019, 0x0017, 0x0016,
+			0x0015, 0x0014, 0x0012, 0x0011, 0x0010, 0x000f, 0x000e, 0x000d
+		];
+
 		const firstRowOfCurrentPattern = $allPatternRows[$currentPatternIndex];
 
 		if (firstRowOfCurrentPattern.isPlaceholder) {
 			return;
 		}
 
+		let noteFreq = PT3ToneTable[0];
+
 		const remainingRows = $allPatternRows.slice(firstRowOfCurrentPattern.globalIndex);
 		let speedDecimal = 3;
 
 		for (const visibleRow of remainingRows) {
 			let speedHex: string | null = null;
+			let noteData: NoteData = new NoteData(Note.None, 0);
 
 			//assuming C > B > A priority
 			if (visibleRow.row.channelsData[2].effect === 'B') {
@@ -90,6 +108,13 @@
 
 			if (speedHex !== null) {
 				speedDecimal = parseInt(speedHex, 16);
+			}
+
+			noteData = visibleRow.row.channelsData[0].noteData;
+			const noteIntValue = noteData.getNoteValue();
+
+			if (noteIntValue) {
+				noteFreq = PT3ToneTable[noteIntValue];
 			}
 
 			const delay = speedDecimal * (1.0 / 50) * 1000;
